Move TaxReturnViewModel (#463)
