+++ app/controllers/actions/CommonAuth.scala
@@ -17,7 +17,6 @@
 package controllers.actions
 
 import config.FrontendAppConfig
-import play.api.mvc.{Result, Results}
 import uk.gov.hmrc.auth.core.retrieve.v2.Retrievals._
 
 trait CommonAuth {
