+++ app/controllers/changeGroupLead/NewGroupLeadConfirmationController.scala
@@ -16,7 +16,6 @@
 
 package controllers.changeGroupLead
 
-import navigation.ChangeGroupLeadNavigator
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
